Fix profiler port (#90)
Two endpoints should be on the profiler port rather than the main
port: `gc_stats` and `live_profiler`. This is done by installing them
on the default mux used by the profiler, rather than a custom mux
passed on to the main server.

Once `live_profiler` is on the internal profiler port, we can remove
the `127.0.0.1` remote-ip restriction to make it easier to use, given
that the default base image is a no-login image.